- Fix aggiunta stazion sismica che dava errore - Inizio realizzazione metodi per inserimento di sensori ed acquisitori

+++ webApp/flask_backend/Model/Componenti/Batteria.py
@@ -4,5 +4,5 @@ from decimal import Decimal
 
 class Batteria(db.Entity):
     componente = PrimaryKey('Componente', column='cod_componente')
-    voltaggio = Required(Decimal, precision=2)
-    amperaggio = Required(Decimal, precision=2)
+    voltaggio = Required(int)
+    amperaggio = Required(int)
