+++ webApp/flask_backend/Model/Componenti/PannelloSolare.py
@@ -5,5 +5,5 @@ from decimal import Decimal
 
 class Pannello_Solare(db.Entity):
     componente = PrimaryKey('Componente', column='cod_componente')
-    potenza_pannello = Optional(Decimal, precision=2)
-    ah_pannello = Required(Decimal, precision=2)  # corrente erogata in AH
+    potenza_pannello = Optional(int) # in Watt
+    ah_pannello = Required(int)  # corrente erogata in AH
