+++ webApp/flask_backend/Model/Componenti/RegolatoreCarica.py
@@ -1,9 +1,9 @@
 from pony.orm import *
 from connection import db
-from decimal import Decimal
+
 
 
 class Regolatore_Carica(db.Entity):
     componente = PrimaryKey('Componente', column='cod_componente')
-    volts_supportati = Required(Decimal, precision=2)
+    volts_supportati = Required(int) # Volt
     ah_supportati = Optional(str, 100)
